Support custom JSON mappers
This adds support for custom JSON mappers allowing mappers to be
provided as needed without breaking existing implementations.

References #37
