+++ lib/protobuf.dart
@@ -27,7 +27,9 @@ part 'src/protobuf/field_set.dart';
 part 'src/protobuf/field_type.dart';
 part 'src/protobuf/generated_message.dart';
 part 'src/protobuf/generated_service.dart';
-part 'src/protobuf/json.dart';
+part 'src/protobuf/json_mapper.dart';
+part 'src/protobuf/json_name_mapper.dart';
+part 'src/protobuf/json_tag_mapper.dart';
 part 'src/protobuf/pb_list.dart';
 part 'src/protobuf/protobuf_enum.dart';
 part 'src/protobuf/readonly_message.dart';
